i2c: i2c-qcom-geni: Change the high and low time for SCL
This change is to meet the spec for i2c at 400KHz. Note that there
is a QUP FW change along with this and then only this will work.
FW change takes care of the duty cycle, clock on/off.

Associated FW change is good to have for this SW change.
FW change CR :  TBD separately by HW team.
CR fixed : 3017809.

Change-Id: I2446b736ef5f6cfa5310f4a30a717272d433a655
Signed-off-by: Mukesh Kumar Savaliya <msavaliy@codeaurora.org>